API Reference

This section provides a detailed reference for all functions and types defined in the AbstractMediumProperties module.

AbstractMediumProperties.DIPPR105Params

AbstractMediumProperties.DIPPR105

AbstractMediumProperties.absorption_length

AbstractMediumProperties.calc_quan_fry_params

AbstractMediumProperties.cherenkov_angle

AbstractMediumProperties.dispersion

AbstractMediumProperties.dispersion_fry

AbstractMediumProperties.es_scattering

AbstractMediumProperties.es_scattering_cumulative

AbstractMediumProperties.es_scattering_integral

AbstractMediumProperties.generic_scattering_function

AbstractMediumProperties.group_refractive_index

AbstractMediumProperties.group_velocity

AbstractMediumProperties.hg_scattering_func

AbstractMediumProperties.make_inverse_es_polynomial

AbstractMediumProperties.mixed_hg_generic_scattering_func

AbstractMediumProperties.mixed_hg_sl_scattering_func

AbstractMediumProperties.mixed_hg_sl_scattering_func_ppc

AbstractMediumProperties.phase_refractive_index

AbstractMediumProperties.pressure

AbstractMediumProperties.refractive_index_fry

AbstractMediumProperties.sca_len_part_conc

AbstractMediumProperties.scattering_function

AbstractMediumProperties.scattering_length

AbstractMediumProperties.sl_scattering_func

AbstractMediumProperties.DIPPR105Params — Type

DIPPR105Params

Parameters for the DIPPR105 formula

AbstractMediumProperties.DIPPR105 — Function

DIPPR105(temperature::Real, params::DIPPR105Params=DDBDIPR105Params)

Use DPPIR105 formula to calculate water density as function of temperature. temperature in K.

Reference: http://ddbonline.ddbst.de/DIPPR105DensityCalculation/DIPPR105CalculationCGI.exe?component=Water

Returns density in kg/m^3

AbstractMediumProperties.absorption_length — Method

absorption_length(wavelength, medium::MediumProperties)

Return absorption length at wavelength in units m.

wavelength is expected to be in units nm.

AbstractMediumProperties.calc_quan_fry_params — Method

calc_quan_fry_params(salinity::Real, temperature::Real, pressure::Real)

Helper function to get the parameters for the Quan & Fry formula as function of salinity, temperature and pressure.

AbstractMediumProperties.cherenkov_angle — Method

cherenkov_angle(wavelength, medium::MediumProperties)

Calculate the cherenkov angle (in rad) for wavelength.

wavelength is expected to be in units nm.

AbstractMediumProperties.dispersion — Method

dispersion(wavelength, medium)

Return the dispersion dn/dλ at wavelength in units 1/nm.

wavelength is expected to be in units nm.

AbstractMediumProperties.dispersion_fry — Method

dispersion_fry(
    wavelength::T;
    salinity::Real,
    temperature::Real,
    pressure::Real) where {T <: Real}

Calculate the dispersion (dn/dλ) for the Quan & Fry model. Wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

AbstractMediumProperties.es_scattering — Method

es_scattering(cos_theta::T, b::T) where {T<:Real}

Einstein-Smoluchowsky Scattering PDF.

AbstractMediumProperties.es_scattering_cumulative — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Integral of ES scattering function from -1 to cos_theta

AbstractMediumProperties.es_scattering_integral — Method

es_scattering_integral(cos_theta::T, b::T) where {T<:Real}

Anti-derivative of the ES scattering function.

AbstractMediumProperties.generic_scattering_function — Method

generic_scattering_function(p0::T, p1::T, p2::T, p3::T) where {T<:Real}

Generic scattering function that can be used to sample from any 3rd order polynomial.

Returns cos(scattering_angle).

AbstractMediumProperties.group_refractive_index — Method

group_refractive_index(wavelength, medium)

Return the group refractive index at wavelength.

wavelength is expected to be in units nm.

AbstractMediumProperties.group_velocity — Method

group_velocity(wavelength, medium)

Return the group_velocity in m/ns at wavelength.

wavelength is expected to be in units nm.

AbstractMediumProperties.hg_scattering_func — Method

hg_scattering_func(g::Real)

CUDA-optimized version of Henyey-Greenstein scattering in one plane.

Arguments

g::Real: mean scattering angle

Returns

typeof(g) cosine of a scattering angle sampled from the distribution

AbstractMediumProperties.make_inverse_es_polynomial — Method

make_inverse_es_polynomial(b::T) where {T<:Real}

Make a 3rd order polynomial that approximates the inverse of the ES scattering function.

AbstractMediumProperties.mixed_hg_generic_scattering_func — Method

mixed_hg_es_scattering_func(g::Real, hg_fraction::Real, poly_coeffs::NTuple{4, Real})

Mixture model of HG and generic scattering function.

AbstractMediumProperties.mixed_hg_sl_scattering_func — Method

mixed_hg_sl_scattering_func(g, hg_fraction)

Mixture model of HG and SL.

Arguments

g::Real: mean scattering angle

hg_fraction::Real: mixture weight of the HG component

AbstractMediumProperties.mixed_hg_sl_scattering_func_ppc — Method

mixed_hg_sl_scattering_func_ppc(g, hg_fraction)

Mixture model of HG and SL as implemented in PPC

AbstractMediumProperties.phase_refractive_index — Method

phase_refractive_index(wavelength, medium)

Return the phase refractive index at wavelength.

wavelength is expected to be in units nm.

AbstractMediumProperties.pressure — Method

pressure(::MediumProperties)

This function returns the pressure for a given medium.

AbstractMediumProperties.refractive_index_fry — Method

refractive_index_fry(wavelength, salinity, temperature, pressure)

The phase refractive index of sea water according to a model from Quan & Fry.

wavelength is given in nm, salinity in permille, temperature in °C and pressure in atm

The original model is taken from: X. Quan, E.S. Fry, Appl. Opt., 34, 18 (1995) 3477-3480.

An additional term describing pressure dependence was included according to: Wolfgang H.W.A. Schuster, "Measurement of the Optical Properties of the Deep Mediterranean - the ANTARES Detector Medium.", PhD thesis (2002), St. Catherine's College, Oxford downloaded Jan 2011 from: http://www.physics.ox.ac.uk/Users/schuster/thesis0098mmjhuyynh/thesis.ps

Adapted from clsim (©Claudio Kopper)

AbstractMediumProperties.sca_len_part_conc — Method

sca_len_part_conc(wavelength; vol_conc_small_part, vol_conc_large_part)

Calculates the scattering length (in m) for a given wavelength based on concentrations of small (vol_conc_small_part) and large (vol_conc_large_part) particles. wavelength is given in nm, volconcsmallpart and volconclargepart in ppm

Adapted from clsim ©Claudio Kopper

AbstractMediumProperties.scattering_function — Method

scattering_function(medium::MediumProperties)

Return a scattering angle sampled from the scattering function of the medium.

AbstractMediumProperties.scattering_length — Method

scattering_length(wavelength, medium::MediumProperties)

Return scattering length at wavelength in units m.

wavelength is expected to be in units nm. Returned length is in units m.

AbstractMediumProperties.sl_scattering_func — Method

sl_scattering_func(g::Real)

Simplified-Liu scattering angle function. Implementation from: https://user-web.icecube.wisc.edu/~dima/work/WISC/ppc/spice/new/paper/a.pdf

Arguments

g::Real: mean scattering angle